Extract LibrariesCommander.ts
