+++ client/TrackerViewModel.ts
@@ -5,7 +5,8 @@ import { Combatant } from "./Combatant/Combatant";
 import { CombatantViewModel } from "./Combatant/CombatantViewModel";
 import { CombatantCommander } from "./Commands/CombatantCommander";
 import { BuildEncounterCommandList } from "./Commands/Command";
-import { EncounterCommander, LibrariesCommander } from "./Commands/EncounterCommander";
+import { EncounterCommander } from "./Commands/EncounterCommander";
+import { LibrariesCommander } from "./Commands/LibrariesCommander";
 import { PromptQueue } from "./Commands/Prompts/PromptQueue";
 import { Toolbar } from "./Commands/components/Toolbar";
 import { Encounter } from "./Encounter/Encounter";
